+++ src/main/java/com/infinityraider/agricraft/blocks/pad/AbstractBlockWaterPad.java
@@ -18,9 +18,6 @@ import net.minecraft.item.ItemStack;
 import net.minecraft.util.math.AxisAlignedBB;
 import com.agricraft.agricore.core.AgriCore;
 import com.infinityraider.infinitylib.render.block.RenderBlock;
-import net.minecraft.block.state.IBlockState;
-import net.minecraft.util.math.BlockPos;
-import net.minecraft.world.World;
 import net.minecraftforge.fml.relauncher.Side;
 import net.minecraftforge.fml.relauncher.SideOnly;
